feat: implement job filters and refactor job naming
- Implement filters for job listings

- Integrate validators for filter parameters

- Resolve bugs in job posting workflow

- Rename 'job' files to 'jobs' for better convention

+++ src/controllers/jobs.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
-import JobService from '../services/job.service.js';
+import JobService from '../services/jobs.service.js';
 import { StatusCodes } from 'http-status-codes';
-import type { CreateJobBody } from '../schemas/job.schema.js';
+import type { CreateJobBody } from '../schemas/jobs.schema.js';
 import { UUID } from 'node:crypto';
 import { sendError, sendSuccess } from '../utils/responseFormatter.js';
 
@@ -22,34 +22,36 @@ class JobController {
   }
 
   async getAllJobsOfEmp(req: Request, res: Response) {
-    const empId = <UUID>req.user?.id || '';
+    if (!req.user) return sendError(res, 'Login first', StatusCodes.BAD_REQUEST);
+    const empId = <UUID>req.user.id;
     const jobs = await this.jobService.getAllJobsOfEmp(empId);
     return sendSuccess(res, jobs, 'Successfully fetched your all posted jobs');
   }
 
   async updateJobStatus(req: Request, res: Response) {
+    if (!req.user) return sendError(res, 'Login first', StatusCodes.BAD_REQUEST);
+    const empId = <UUID>req.user.id;
     const jobId = <UUID>req.params.id;
-    const empId = <UUID>req.user?.id || '';
     const updatedJob = await this.jobService.updateJobStatus(jobId, empId);
     return sendSuccess(res, updatedJob, 'Successfully job status updated', StatusCodes.OK);
   }
 
   async deleteJobById(req: Request, res: Response) {
+    if (!req.user) return sendError(res, 'Login first', StatusCodes.BAD_REQUEST);
+    const empId = <UUID>req.user.id;
     const jobId = <UUID>req.params.id;
-    const empId = <UUID>req.user?.id;
     const response = await this.jobService.deleteJobById(jobId, empId);
     return sendSuccess(res, { jobDeleted: response }, 'Successfully deleted');
   }
 
   async getAllJobs(req: Request, res: Response) {
-    const jobs = await this.jobService.getAllJobs();
+    const jobs = await this.jobService.getAllJobs(req.query);
     return sendSuccess(res, jobs, 'Succssfully fetched all jobs', StatusCodes.OK);
   }
 
   async getJobById(req: Request, res: Response) {
     const jobId = <UUID>req.params.id;
     const response = await this.jobService.getJobById(jobId);
-    if (typeof response === 'string') return sendError(res, response, StatusCodes.NOT_FOUND);
     const skills = response.skills.map(({ skill }) => ({ id: skill.id, name: skill.name }));
     const job = { ...response, skills: skills };
     return sendSuccess(res, job, 'Successfully fetched job details');
